+++ app/collection/[id]/actions/actions.ts
@@ -1,3 +1,18 @@
+export interface Transaction {
+  id: string;
+  type: 'PAYMENT' | 'WITHDRAWAL';
+  parent: string;
+  child: string;
+  amount: number;
+  date: string;
+  rawDate: Date;
+}
+
+export interface AttachmentProps {
+  id: string;
+  label: string;
+}
+
 export const changeCollectionCover = () => {};
 
 export const updateCollectionDescription = (newDescription: string) => {};
