+++ app/collection/[id]/components/CollectionTitleCard.tsx
@@ -18,6 +18,7 @@ import AddIcon from '@mui/icons-material/Add';
 import CheckIcon from '@mui/icons-material/Check';
 import CloseIcon from '@mui/icons-material/Close';
 import {
+  AttachmentProps,
   changeCollectionCover,
   deleteAttachment,
   downloadAttachment,
@@ -26,7 +27,6 @@ import {
   uploadAttachment,
 } from '../actions/actions';
 import { useState, useTransition } from 'react';
-import { AttachmentProps } from '../page';
 
 interface CollectionTitleCardProps {
   coverImage: string | null;
